requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
boto3==1.19.2
botocore==1.22.2
Django==3.2.8
django-cors-headers==3.10.0
django-storages==1.12.2
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
promise==2.3
psycopg2==2.9.1
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
Rx==1.6.1
s3transfer==0.5.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
whitenoise==5.3.0
